refactor(addin): render the Word add-in's issues with the shared row and drop the old card
The add-in was the last screen on DocumentIssueCard. It now uses the
explorer's IssuesList: grouped by severity, one row open at a time, and
opening a row (or its "Select in document" link) selects the paragraph in
Word. The list is read-only for resolving, since the pane runs on a share
token with no session to act as.

- IssuesList takes a crossLink render prop for hosts without the project
  view context
- IssueDocumentLink takes a label, since a line number means little in Word
- IssueFeedbackButtons and feedbackLabel move to their own module
- Delete document-issue-card.tsx and document-issues-list.tsx
